refactor(semantic): split command topology module
